+++ sys/net/application_layer/telnet/telnet_server.c
@@ -293,7 +293,7 @@ disco:
     return NULL;
 }
 
-ssize_t telnet_server_write(const void* buffer, size_t len)
+int telnet_server_write(const void* buffer, size_t len)
 {
     if (connected) {
         int res = _write_buffer(buffer, len);
